# docs/swagger.yaml
definitions:
...
  entity.Tag:
    properties:
      color:
        type: string
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
...
  request.CreateStatusPageRequest:
    properties:
      description:
        type: string
      monitor_ids:
        items:
          type: integer
        type: array
      name:
        type: string
      slug:
        type: string
    required:
    - name
    - slug
    type: object
...
  request.UpdateStatusPageRequest:
    properties:
      description:
        type: string
      monitor_ids:
        items:
          type: integer
        type: array
      name:
        type: string
      slug:
        type: string
    required:
    - name
    - slug
    type: object
...
  response.MonitorResponse:
    properties:
      check_ssl:
        type: boolean
      created_at:
        type: string
      heartbeat_token:
        type: string
      id:
        type: integer
      interval:
        type: integer
      last_checked_at:
        type: string
      last_latency:
        type: integer
      name:
        type: string
      ssl_expires_at:
        type: string
      status:
        type: string
      tags:
        items:
          $ref: '#/definitions/entity.Tag'
        type: array
      timeout:
        type: integer
      type:
        type: string
      updated_at:
        type: string
      uptime_status:
        type: string
      url:
        type: string
      user_id:
        type: integer
    type: object
  response.StatusPageResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      monitors:
        items:
          $ref: '#/definitions/response.MonitorResponse'
        type: array
      name:
        type: string
      slug:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
...
  utils.ErrorResponseModel:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
    type: object
  utils.SuccessResponseModel:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /api/status-pages:
    get:
      consumes:
      - application/json
      description: Get all status pages for the authenticated user (admin sees all)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.SuccessResponseModel'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/response.StatusPageResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
      security:
      - BearerAuth: []
      summary: List status pages
      tags:
      - Status Pages
    post:
      consumes:
      - application/json
      description: Create a new status page with selected monitors
      parameters:
      - description: Status page details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.CreateStatusPageRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/utils.SuccessResponseModel'
            - properties:
                data:
                  $ref: '#/definitions/response.StatusPageResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
      security:
      - BearerAuth: []
      summary: Create a status page
      tags:
      - Status Pages
  /api/status-pages/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a status page by ID
      parameters:
      - description: Status page ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
      security:
      - BearerAuth: []
      summary: Delete a status page
      tags:
      - Status Pages
    get:
      consumes:
      - application/json
      description: Retrieve a single status page by its ID
      parameters:
      - description: Status page ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.SuccessResponseModel'
            - properties:
                data:
                  $ref: '#/definitions/response.StatusPageResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
      security:
      - BearerAuth: []
      summary: Get status page by ID
      tags:
      - Status Pages
    put:
      consumes:
      - application/json
      description: Update an existing status page
      parameters:
      - description: Status page ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated status page details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.UpdateStatusPageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.SuccessResponseModel'
            - properties:
                data:
                  $ref: '#/definitions/response.StatusPageResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
      security:
      - BearerAuth: []
      summary: Update a status page
      tags:
      - Status Pages
...
  /api/status-pages/{slug}/badge.svg:
    get:
      description: Get a Shields.io-style SVG badge for a status page
      parameters:
      - description: Status page slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/xml
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Get status badge SVG
      tags:
      - Status Pages
  /api/status-pages/slug/{slug}:
    get:
      consumes:
      - application/json
      description: Retrieve a status page publicly by its slug
      parameters:
      - description: Status page slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.SuccessResponseModel'
            - properties:
                data:
                  $ref: '#/definitions/response.StatusPageResponse'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
      summary: Get public status page by slug
      tags:
      - Status Pages
...
swagger: "2.0"